Added function  "NormDoubleRound" for norms double values mantissa to fixed digits(13) Norms sy.B(13), tp.G(13), bcp->C(6) Changed all work arrays COMPOS from float to double Changed in s_formula.cpp(.h) and near internal values from float to double Changed mtp->Ae, mtp->An, usp->A object types from float to double Changed sizes of edit for some objects on DComp window Changed gemipm and tnode examples as in documentation SD
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@1373 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/m_compos.h
@@ -72,7 +72,7 @@ typedef struct
     char (*SA)[MAXFORMUNIT];// List of formula units
     char (*SM)[DC_RKLEN]; // List of DC record keys      [0:Ld-1]
     // COMPOS Work arrays
-    float *A,   // stoichiometry matrix formul. ed and DC  [0:Nmax-1]
+    double *A,   // stoichiometry matrix formul. ed and DC  [0:Nmax-1]
     *ICw  //IC atomic (molar) masses [0:Nmax-1]
     ;
     char (*SB1)[IC_RKLEN], //Complete work list of IC keys in ICOMP chain [0:Nmax-1]
@@ -93,7 +93,7 @@ class TCompos : public TCModule
 
     // work data for calc
     double  *C;
-    float *CI;
+    double *CI;
     char *CIcl;
